Rename driver to pico_board/pico-board
